Consistently handle TimeLineID as UInt32 everywhere (#5101)
TimeLineID is an uint32 in PostgreSQL but since PostgreSQL's
type system is lacking unsigned types there was no consistent
mode of transmission in the streaming replication protocol and
the timeline got transmitted both as integer and as bigint.
As a consequence although we realized the TimeLineID's unsignedness
we also had heterogenity regarding timeline id's ocurrences across
Npgsql's replication API.
AS of PG 16 the backend is consistently transmitting TimeLineID as
bigint that (due to being backed by an uint32 field) can only have
values >= 0.
With this commit we accept TimeLineID as both integer and bigint
and expose it as uint (UInt32) in all our public APIs.

Closes #5046

+++ src/Npgsql/Replication/ReplicationSystemIdentification.cs
@@ -7,7 +7,7 @@ namespace Npgsql.Replication;
 /// </summary>
 public class ReplicationSystemIdentification
 {
-    internal ReplicationSystemIdentification(string systemId, ulong timeline, NpgsqlLogSequenceNumber xLogPos, string dbName)
+    internal ReplicationSystemIdentification(string systemId, uint timeline, NpgsqlLogSequenceNumber xLogPos, string dbName)
     {
         SystemId = systemId;
         Timeline = timeline;
@@ -24,7 +24,7 @@ public class ReplicationSystemIdentification
     /// <summary>
     /// Current timeline ID. Also useful to check that the standby is consistent with the master.
     /// </summary>
-    public ulong Timeline { get; }
+    public uint Timeline { get; }
 
     /// <summary>
     /// Current WAL flush location. Useful to get a known location in the write-ahead log where streaming can start.
